Add files via upload
This project is one of the tasks required in the course "Object Oriented Programming in C# and the .NET Framework". This project is a Cows and Bulls game for a Windows interface.

We've separated the logical part from the interface part, so we can use it again when we will want to replace inetrface for the game. The code written as readable as possible.